+++ plugins/sdk/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/sdk.module';
-export * from './lib/services/navigation.service';
-export * from './lib/services/preview.service';
+
+export * from './lib/components';
+export * from './lib/services';