+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `Preview Service` in the `@app/sdk` library to provide reusable Viewer APIs
 - Support icon buttons with Header component
 
+### Changed
+
+- Move `AppHeaderComponent` component to `@app/sdk`
+
 ## [3.0.0] - 2024-03-12
 
 ### Added